pyproject.toml
[project]
name = "ai-agent-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "authlib>=1.6.0",
    "bcrypt==4.0.1",
    "duckduckgo-search>=8.0.4",
    "fastapi>=0.115.14",
    "gunicorn>=23.0.0",
    "huggingface-hub>=0.33.1",
    "itsdangerous>=2.2.0",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.0",
    "uvicorn>=0.35.0",
]
